Don't check movie count when setting single value, just pass source setting Update database cleaner to check for missing episodes if enabled Move recursive check to cleaner

+++ EmberAPI/clsAPIDatabase.vb
@@ -1743,6 +1743,18 @@ Public Class Database
         Try
             Using SQLtransaction As SQLite.SQLiteTransaction = SQLcn.BeginTransaction
                 If CleanMovies Then
+
+                    'get a listing of sources and their recursive properties
+                    Dim sDict As New Dictionary(Of String, Boolean)
+                    Using SQLcommand As SQLite.SQLiteCommand = Master.DB.CreateCommand
+                        SQLcommand.CommandText = "SELECT Path, Recursive FROM sources;"
+                        Using SQLreader As SQLite.SQLiteDataReader = SQLcommand.ExecuteReader()
+                            While SQLreader.Read
+                                sDict.Add(SQLreader("Path").ToString, Convert.ToBoolean(SQLreader("Recursive")))
+                            End While
+                        End Using
+                    End Using
+
                     Using SQLcommand As SQLite.SQLiteCommand = SQLcn.CreateCommand
                         SQLcommand.CommandText = "SELECT MoviePath, Id FROM movies ORDER BY ListTitle COLLATE NOCASE;"
                         Using SQLReader As SQLite.SQLiteDataReader = SQLcommand.ExecuteReader()
@@ -1754,6 +1766,8 @@ Public Class Database
                                     If ((Not Master.eSettings.SkipStackSizeCheck OrElse Not StringUtils.IsStacked(fInfo.Name)) AndAlso fInfo.Length < Master.eSettings.SkipLessThan * 1048576) Then
                                         Me.DeleteFromDB(Convert.ToInt64(SQLReader("ID")), True)
                                     End If
+                                ElseIf sDict.FirstOrDefault(Function(s) SQLReader("MoviePath").ToString.ToLower.StartsWith(s.Key.ToLower)).Value = False AndAlso Convert.ToBoolean(SQLReader("Recursive")) Then
+                                    Me.DeleteFromDB(Convert.ToInt64(SQLReader("ID")), True)
                                 End If
                             End While
                         End Using
@@ -1854,6 +1868,8 @@ Public Class Database
         Try
             Dim tID As Integer = -1
             Dim tSeason As Integer = -999
+            Dim tMissing As Boolean = False
+
             Dim SQLtransaction As SQLite.SQLiteTransaction = Nothing
             If Not BatchMode Then SQLtransaction = SQLcn.BeginTransaction
             Using SQLcommand As SQLite.SQLiteCommand = SQLcn.CreateCommand
@@ -1861,36 +1877,44 @@ Public Class Database
                 Using SQLreader As SQLite.SQLiteDataReader = SQLcommand.ExecuteReader
                     While SQLreader.Read
                         Using SQLECommand As SQLite.SQLiteCommand = SQLcn.CreateCommand
-                            SQLECommand.CommandText = String.Concat("SELECT TVShowID, Season FROM TVEps WHERE TVEpPathID = ", SQLreader("ID"), ";")
+                            SQLECommand.CommandText = String.Concat("SELECT TVShowID, Season, Missing FROM TVEps WHERE TVEpPathID = ", SQLreader("ID"), ";")
                             Using SQLEReader As SQLite.SQLiteDataReader = SQLECommand.ExecuteReader
                                 While SQLEReader.Read
                                     tID = Convert.ToInt32(SQLEReader("TVShowID"))
                                     tSeason = Convert.ToInt32(SQLEReader("Season"))
+                                    tMissing = Convert.ToBoolean(SQLEReader("Missing"))
                                 End While
                             End Using
                         End Using
 
-                        SQLcommand.CommandText = String.Concat("DELETE FROM TVEps WHERE TVEpPathID = ", SQLreader("ID"), ";")
-                        SQLcommand.ExecuteNonQuery()
-                        SQLcommand.CommandText = String.Concat("DELETE FROM TVEpActors WHERE TVEpID = ", SQLreader("ID"), ";")
-                        SQLcommand.ExecuteNonQuery()
-                        SQLcommand.CommandText = String.Concat("DELETE FROM TVVStreams WHERE TVEpID = ", SQLreader("ID"), ";")
-                        SQLcommand.ExecuteNonQuery()
-                        SQLcommand.CommandText = String.Concat("DELETE FROM TVAStreams WHERE TVEpID = ", SQLreader("ID"), ";")
-                        SQLcommand.ExecuteNonQuery()
-                        SQLcommand.CommandText = String.Concat("DELETE FROM TVSubs WHERE TVEpID = ", SQLreader("ID"), ";")
-                        SQLcommand.ExecuteNonQuery()
+                        If Not tMissing Then
+                            If Not Master.eSettings.DisplayMissingEpisodes Then
+                                SQLcommand.CommandText = String.Concat("DELETE FROM TVEps WHERE TVEpPathID = ", SQLreader("ID"), ";")
+                                SQLcommand.ExecuteNonQuery()
+                                SQLcommand.CommandText = String.Concat("DELETE FROM TVEpActors WHERE TVEpID = ", SQLreader("ID"), ";")
+                                SQLcommand.ExecuteNonQuery()
+                                SQLcommand.CommandText = String.Concat("DELETE FROM TVVStreams WHERE TVEpID = ", SQLreader("ID"), ";")
+                                SQLcommand.ExecuteNonQuery()
+                                SQLcommand.CommandText = String.Concat("DELETE FROM TVAStreams WHERE TVEpID = ", SQLreader("ID"), ";")
+                                SQLcommand.ExecuteNonQuery()
+                                SQLcommand.CommandText = String.Concat("DELETE FROM TVSubs WHERE TVEpID = ", SQLreader("ID"), ";")
+                                SQLcommand.ExecuteNonQuery()
 
-                        SQLcommand.CommandText = String.Concat("SELECT ID FROM TVEps WHERE TVShowID = ", tID, " AND Season = ", tSeason, ";")
-                        Using SQLSeasonReader As SQLite.SQLiteDataReader = SQLcommand.ExecuteReader
-                            If Not SQLSeasonReader.HasRows Then
-                                'no more episodes for this season, delete the season
-                                Using SQLSeasonCommand As SQLite.SQLiteCommand = SQLcn.CreateCommand
-                                    SQLSeasonCommand.CommandText = String.Concat("DELETE FROM TVSeason WHERE TVShowID = ", tID, " AND Season = ", tSeason, ";")
-                                    SQLSeasonCommand.ExecuteNonQuery()
+                                SQLcommand.CommandText = String.Concat("SELECT ID FROM TVEps WHERE TVShowID = ", tID, " AND Season = ", tSeason, ";")
+                                Using SQLSeasonReader As SQLite.SQLiteDataReader = SQLcommand.ExecuteReader
+                                    If Not SQLSeasonReader.HasRows Then
+                                        'no more episodes for this season, delete the season
+                                        Using SQLSeasonCommand As SQLite.SQLiteCommand = SQLcn.CreateCommand
+                                            SQLSeasonCommand.CommandText = String.Concat("DELETE FROM TVSeason WHERE TVShowID = ", tID, " AND Season = ", tSeason, ";")
+                                            SQLSeasonCommand.ExecuteNonQuery()
+                                        End Using
+                                    End If
                                 End Using
+                            Else
+                                SQLcommand.CommandText = String.Concat("UPDATE TVEps SET Missing = 1 WHERE TVEpID = ", SQLreader("ID"), ";")
+                                SQLcommand.ExecuteNonQuery()
                             End If
-                        End Using
+                        End If
                     End While
                 End Using
             End Using
